+++ ConcordiumWalletTests/Views/AccountsView/AccountDetails/TransactionsLoadingHandlerTests.swift
@@ -5,79 +5,11 @@
 
 import Foundation
 import XCTest
-import Combine
-
 @testable import Mock
-
-class DependencyProviderMock: DependencyProviderMockHelper {
-    var transactionsServiceMock: TransSrcvMock
-    var storageManagerMock: StorageManagerMock
-
-    init(transactionsServiceMock: TransSrcvMock, storageManagerMock: StorageManagerMock) {
-        self.transactionsServiceMock = transactionsServiceMock
-        self.storageManagerMock = storageManagerMock
-        super.init()
-    }
-
-    override func transactionsService() -> TransactionsServiceProtocol {
-        transactionsServiceMock
-    }
-
-    override func storageManager() -> StorageManagerProtocol {
-        storageManagerMock
-    }
-}
-
-class StorageManagerMock: StorageManagerMockHelper {
-    var transfers = [TransferDataType]()
-
-    override func getRecipient(withAddress address: String) -> RecipientDataType? {
-        RecipientEntity()
-    }
-
-    override func getTransfers(for accountAddress: String) -> [TransferDataType] {
-        transfers
-    }
-
-    func addMockLocalTransaction(time: Int) {
-        let transferEntity = TransferEntity()
-        transferEntity.createdAt = Date(timeIntervalSince1970: Double(time))
-        transfers.append(transferEntity)
-    }
-}
-
-class TransSrcvMock: TransactionsServiceMockHelper {
-    var transactions = [Transaction]()
-    var limit = 10
-
-    override func getTransactions(for account: AccountDataType, startingFrom: Transaction?) -> Combine.AnyPublisher<RemoteTransactions, Error> {
-        var dropFirst = 0
-        if let startingFrom = startingFrom {
-            let foundIndex = transactions.firstIndex { transaction in transaction.blockTime == startingFrom.blockTime }
-            dropFirst = foundIndex != nil ? foundIndex! + 1 : 0
-        }
-        let transactionsToReturn = Array(transactions.dropFirst(dropFirst).prefix(limit))
-        return .just(RemoteTransactions(transactions: transactionsToReturn, count: transactionsToReturn.count, limit: limit, order: "d"))
-    }
-
-    func addMockRemoteTransaction(time: Double) {
-        
-        let mockDetails: Details = Details(transferDestination: nil, memo: nil, transferAmount: nil, events: nil,
-                                           outcome: .success, type: nil, detailsDescription: nil, transferSource: nil,
-                                           newIndex: nil, inputEncryptedAmount: nil, newSelfEncryptedAmount: nil,
-                                           encryptedAmount: nil, aggregatedIndex: nil, amountSubtracted: nil,
-                                           rejectReason: nil, amountAdded: nil)
-        transactions.append(Transaction(blockTime: time, origin: nil, energy: nil, blockHash: "",
-                cost: "0", subtotal: "0",
-                transactionHash: "", details: mockDetails,
-                total: "0", id: 0, encrypted: nil))
-        transactions.sort { $0.blockTime ?? 0 > $1.blockTime ?? 0 } // sort descending
-    }
-}
 
 // swiftlint:disable:next type_body_length
 class TransactionsLoadingHandlerTests: XCTestCase {
-    let transactionsServiceMock = TransSrcvMock()
+    let transactionsServiceMock = TransServiceMock()
     let storageManagerMock = StorageManagerMock()
     var dp: DependencyProviderMock!
     var account = AccountEntity()
